+++ src/domain/application/layout/kukkuuDetailPage/KukkuuDetailPage.tsx
@@ -1,5 +1,5 @@
 import React, { ComponentType, ReactElement } from 'react';
-import { ResourceComponentPropsWithId, useGetOne, Record } from 'react-admin';
+import { useGetOne, Record, ShowProps } from 'react-admin';
 import omit from 'lodash/omit';
 
 import { Crumb } from '../../../../common/components/breadcrumbs/Breadcrumbs';
@@ -9,9 +9,8 @@ import KukkuuShow from './KukkuuShow';
 
 type Props = {
   children: ReactElement;
-  reactAdminProps: ResourceComponentPropsWithId;
-  pageTitleSource: string;
-  actions?: ReactElement;
+  reactAdminProps: ShowProps;
+  pageTitleSource?: string;
   layout?: ComponentType<KukkuuLayoutProps>;
   breadcrumbs?: ((data?: Record) => Crumb[]) | Crumb[];
 };
@@ -20,7 +19,6 @@ const KukkuuDetailPage = ({
   children,
   pageTitleSource,
   reactAdminProps,
-  actions,
   layout: Layout = KukkuuCardPageLayout,
   breadcrumbs,
 }: Props) => {
@@ -40,9 +38,7 @@ const KukkuuDetailPage = ({
       reactAdminProps={reactAdminProps}
       breadcrumbs={crumbs}
     >
-      <KukkuuShow actions={actions} {...omit(reactAdminProps, 'hasShow')}>
-        {children}
-      </KukkuuShow>
+      <KukkuuShow {...omit(reactAdminProps, 'hasShow')}>{children}</KukkuuShow>
     </Layout>
   );
 };